주문 생성 API와 주문 결과 페이지 연동

+++ src/main/java/com/shop/domain/Orders.java
@@ -12,7 +12,7 @@ public class Orders {
 	private String ordererPhoneNumber;
 	private String ordererEmail;
 	private String orderStatus;
-	private Integer totalPrice;
+	private Long totalPrice;
 	private Date createdAt;
 	private Date paidAt;
 	private String receiverName;
